Extract workspace init from base_projects engine
